Connected DataBase to gui :)
